requirements.txt
Flask
python-dotenv
Flask-MySQLdb
MySQL-python
